Fixed more issues
