+++ features/balances/src/main/kotlin/es/pedrazamiguez/splittrip/features/balance/presentation/screen/CategorySpendingScreen.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -21,9 +20,11 @@ import es.pedrazamiguez.splittrip.core.designsystem.navigation.LocalBottomPaddin
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.chart.DonutChart
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.chart.DonutChartData
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.layout.DeferredLoadingContainer
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.layout.FlatCard
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.layout.ShimmerLoadingList
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.topbar.rememberConnectedScrollBehavior
 import es.pedrazamiguez.splittrip.features.balance.R
+import es.pedrazamiguez.splittrip.features.balance.presentation.component.CategorySpendingItemRow
 import es.pedrazamiguez.splittrip.features.balance.presentation.viewmodel.state.CategorySpendingUiState
 import kotlinx.collections.immutable.toImmutableList
 
@@ -54,16 +55,26 @@ fun CategorySpendingScreen(
                 CategorySpendingChart(uiState = uiState)
             }
 
-            items(uiState.items) { item ->
-                CategorySpendingItemRow(
-                    item = item,
+            item {
+                FlatCard(
                     modifier = Modifier
                         .fillMaxWidth()
-                        .padding(
-                            horizontal = MaterialTheme.spacing.Medium,
-                            vertical = MaterialTheme.spacing.Small
-                        )
-                )
+                        .padding(MaterialTheme.spacing.Medium)
+                ) {
+                    Column {
+                        uiState.items.forEach { item ->
+                            CategorySpendingItemRow(
+                                item = item,
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .padding(
+                                        horizontal = MaterialTheme.spacing.Default,
+                                        vertical = MaterialTheme.spacing.Medium
+                                    )
+                            )
+                        }
+                    }
+                }
             }
         }
     }
@@ -83,7 +94,7 @@ private fun CategorySpendingChart(uiState: CategorySpendingUiState) {
         data = chartData,
         modifier = Modifier
             .fillMaxWidth()
-            .padding(horizontal = 64.dp, vertical = MaterialTheme.spacing.Large),
+            .padding(horizontal = 64.dp, vertical = MaterialTheme.spacing.Screen),
         centerContent = {
             if (uiState.totalFormattedAmount.isNotBlank()) {
                 Column(
